RANSAC moved to separate file
